pyproject.toml
[project]
name = "flow-chest"
version = "0.1.0"
description = "Mini orchestrator estilo n8n para aprender low-code + IA"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.116.1",
  "uvicorn==0.35.0",
  "pydantic==2.11.7",
  "requests==2.32.4",
  "python-dotenv==1.1.1"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
